% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend.R
\name{plot_trend}
\alias{plot_trend}
\title{Plot Trend}
\usage{
plot_trend(data, x, y, group = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{x}{Name of the column for x-axis (unquoted).}

\item{y}{Name of the column for y-axis (unquoted).}

\item{group}{Optional column to group data by (unquoted).}
}
\value{
A ggplot object.
}
\description{
Visualizes a trend from a data frame.
}
\examples{
library(dplyr)
df <- data.frame(Day = 1:10, Weight = rnorm(10))
plot_trend(df, Day, Weight)
}
